+++ tap-snapshots/test/root.ts.test.cjs
@@ -1,0 +1,90 @@
+/* IMPORTANT
+ * This snapshot file is auto-generated, but designed for humans.
+ * It should be checked into source control and tracked carefully.
+ * Re-generate by setting TAP_SNAPSHOT=1 and running tests.
+ * Make sure to inspect the output below.  Do not ignore changes!
+ */
+'use strict'
+exports[`test/root.ts TAP set root option absolute=false > async 1`] = `
+Array [
+  "x/x/a",
+  "x/x/x/a",
+  "x/x/x/y",
+  "x/x/x/y/r",
+  "x/x/y",
+  "x/x/y/r",
+  "x/y",
+  "x/y/r",
+  "y/r",
+]
+`
+
+exports[`test/root.ts TAP set root option absolute=false > sync 1`] = `
+Array [
+  "x/x/a",
+  "x/x/x/a",
+  "x/x/x/y",
+  "x/x/x/y/r",
+  "x/x/y",
+  "x/x/y/r",
+  "x/y",
+  "x/y/r",
+  "y/r",
+]
+`
+
+exports[`test/root.ts TAP set root option absolute=true > async 1`] = `
+Array [
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/a",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/x/a",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/x/y",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/x/y/r",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/y",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/y/r",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/y",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/y/r",
+  "{CWD}/test/tap-testdir-root-set-root-option/y/r",
+]
+`
+
+exports[`test/root.ts TAP set root option absolute=true > sync 1`] = `
+Array [
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/a",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/x/a",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/x/y",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/x/y/r",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/y",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/y/r",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/y",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/y/r",
+  "{CWD}/test/tap-testdir-root-set-root-option/y/r",
+]
+`
+
+exports[`test/root.ts TAP set root option absolute=undefined > async 1`] = `
+Array [
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/a",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/x/a",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/x/y",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/y",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/y",
+  "x/x/x/y/r",
+  "x/x/y/r",
+  "x/y/r",
+  "y/r",
+]
+`
+
+exports[`test/root.ts TAP set root option absolute=undefined > sync 1`] = `
+Array [
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/a",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/x/a",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/x/y",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/x/y",
+  "{CWD}/test/tap-testdir-root-set-root-option/x/y",
+  "x/x/x/y/r",
+  "x/x/y/r",
+  "x/y/r",
+  "y/r",
+]
+`